Добавил первоначальную миграцию БД
